requirements.txt
Flask>=3.0
Flask-Login>=0.6
python-dotenv  # load environment variables from .env files
google-api-python-client
google-auth
gunicorn>=20.1.0
